fix: New code formatted
