+++ src/routes/overview-page/+page.svelte
@@ -136,7 +136,8 @@
     $effect(() => {
         if (
             (branch_selection && branch_selection !== "") ||
-            (start_date && end_date)
+            (start_date && end_date) ||
+            (start_date === "" && end_date === "")
         ) {
             (async () => {
                 // Fetch new contributors for the selected branch
